Trusted Host supports multiple keys per host.
Get-SSHTrustedHost has been updated to retrieve fingerprints from the
registry, each value is expected to be a MultiString type value.
Set-SSHTrustedHost has been updated to store fingerprints in the
registry, each new value is added or used to create a new MultiString
type value. It will overwrite old keys with the old value and the new
value.